% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Defines the server logic for the FastLinkGUIne app}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{\itemize{
\item input from UI
}}

\item{output}{\itemize{
\item output to UI
}}

\item{session}{\itemize{
\item user session
}}
}
\description{
This function is the backend for the FastLinkGUIne App.
}
